+++ lib/snacky.dart
@@ -1,4 +1,5 @@
 export 'src/builder/simple_snacky_builder.dart';
+export 'src/builder/toast_snacky_builder.dart';
 export 'src/builder/snacky_builder.dart';
 export 'src/controller/snacky_controller.dart';
 export 'src/model/cancelable_snacky.dart';